Fix OpenMP compilation on gcc 7.4.0
The #pragma omp parallel is implemented by creating another function for the thread to launch.
gcc 7.4.0 fails to carry the target attributes to that new function, so intrinstics were not working.
Copying register types causes an internal compiler error.
These for loops need constants initialized in registers like -127 and just moving the constants into the for loop was generating code that initializes them every iteration (no cross-loop constant extraction).
The workaround is to split #pragma omp parallel to launch a function with target attributes which initializes the constants then does #pragma omp for to just divvy up the work.